requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Flask==3.1.0
Flask-SQLAlchemy==2.5.1
mysqlclient==2.1.0
python-dotenv==0.19.2
flask
flask_sqlalchemy
python-dotenv
mysqlclient